Fixed minor things that led to warnings.

+++ include/VEmu.h
@@ -85,7 +85,7 @@ private:
 
 private:
 	Bus bus;
-	uint32_t get_4byte_aligned_instr(uint32_t);
+	uint32_t get_4byte_aligned_instr(uint64_t);
 
 	uint64_t load(uint64_t, size_t);
 	void store(uint64_t, uint64_t, size_t);
